feat: Erstelle MinimalSidebar mit Unterkomponenten für Navigation und Header
